requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
faiss-cpu
openai
streamlit
requests
pydantic
numpy
pandas